+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 
 - Append mode to multi update data attributes task [#4204]
+- Sticky navbar on browse nomenclature and collecting event
 
 ### Changed